apps/curation/pyproject.toml
[tool.poetry]
name = "superstack"
version = "0.1.0"
description = ""
authors = ["Henry Nguyen <henry2833@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.10"
trafilatura = "^1.6.1"
aiohttp = "^3.8.5"
lmdb = "^1.4.1"
newspaper3k = "^0.2.8"
pytest = "^7.4.0"
autopep8 = "^2.0.2"
argparse = "^1.4.0"
nltk = "^3.8.1"
numpy = "^1.25.2"
validators = "^0.21.2"
beautifulsoup4 = "^4.12.2"
load-dotenv = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

